Add target and rel attributes to relative links
This commit adds target and rel attributes to the relative links found
in tuscan_chicken_soup.html, lasagna.html, and index.html. The target
attribute is set such that users open the links in new tabs, allowing
them to keep the original page open for easier navigation. The rel
attribute is set such that the new page cannot know the user's origin
and cannot modify the original page. This attribute enforces security
and privacy for the user.