Set rel attributes in recipe webpages to default
This commit defaults the rel attribute in the home page links found in
the Lasagna Recipe and Tuscan Chicken Soup Recipe webpages. Doing so
allows the user to navigate back to the home page without having too
many tabs open. As the links to each recipe webpage in the homepage open
in a new tab, the user also likely already has the home page open.